# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: orchid API
  version: 0.1.0
  description: |
    The orchid multi-tenant REST API for engineering leadership assessment.
    Hiring platforms integrate via JWT-authenticated endpoints to provision
    sessions, retrieve orchestration fingerprints, and receive webhook
    notifications on fingerprint assembly.
servers:
  - url: https://api.orchid.example.com
    description: Production
  - url: http://localhost:8000
    description: Local development
security:
  - BearerAuth: []
  - TenantHeader: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    TenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-ID
  schemas:
    CandidateRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          nullable: true
        name:
          type: string
          nullable: true
    CandidateResponse:
      type: object
      required: [candidate_id, tenant_id, created_at]
      properties:
        candidate_id:
          type: string
          format: uuid
        tenant_id:
          type: string
        email:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
paths:
  /candidates:
    post:
      summary: Register a candidate
      operationId: registerCandidate
      tags: [candidates]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CandidateRequest'
            example:
              email: alice@example.com
              name: Alice Chen
      responses:
        '201':
          description: Candidate registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
  /candidates/{candidate_id}:
    get:
      summary: Get a candidate
      operationId: getCandidate
      tags: [candidates]
      parameters:
        - name: candidate_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Candidate found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateResponse'
        '404':
          description: Candidate not found
...